개수 Long에서 Int로 통일

+++ src/main/kotlin/com/wafflytime/post/dto/PostResponse.kt
@@ -40,7 +40,7 @@ data class PostResponse(
                 images = images?.sortedBy { it.imageId },
                 nLikes = postEntity.nLikes,
                 nScraps = postEntity.nScraps,
-                nReplies = postEntity.nReplies.toInt()
+                nReplies = postEntity.nReplies
             )
         }
     }
